+++ src/main/java/com/spring/CourseElection/model/response/info/GradeRes.java
@@ -20,7 +20,7 @@ public class GradeRes {
 
     private Double grade;
 
-    private Double usual;
+    private Integer usual;
 
-    private Double examination;
+    private Integer examination;
 }
